feat: Add support for VAE selection in API
The code changes in `api.py` introduce support for selecting VAE options in the API. If the `vae` parameter is set to "Automatic" or "None", the corresponding value is returned. This change enhances the flexibility of the API by allowing users to specify their preferred VAE option.

Note: This commit message follows the established convention of using a prefix (feat) to indicate a new feature or enhancement.